requirements.txt
numpy
opencv-python
imutils
bump2version
pyyaml
pytest
pytest-cov
sklearn